Add/Update: Credits sequence implemented.
Moved congrats out of outro and into its own file, congrats.py. The
combination of congrats and credits in one module made for too large and
complex a module for my taste.

Main and ui\__init__ also updated accordingly to run credits.

Game is now essentially back to the functionality of the master branch.